app/requirements.txt
peewee==2.8.2
tornado==5.1.1
PyMySQL==0.9.2
redis==2.10.6
structlog
pyyaml